[maxime/link-category] Link advert with category + Refacto login, register, SQL

+++ src/com/lat/beans/Group.java
@@ -1,16 +1,14 @@
 package com.lat.beans;
 
-
 public class Group {
-
-    private Long id;
+    private Integer id;
     private String name;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
